dbm: Fix render node usage for wlroots
At least wlroots uses the render nodes for everything possible.
This avoids many issues related to using the primary nodes as described
in the wlroots commits for render/allocator/allocator.c.

As dbm allocates buffers for the blobs, we must use the primary DRM node
for allocation instead of the render node. To do this, let's always open
also the primary node on dbm device creation and close it when the device
gets destroyed.

Signed-off-by: Tony Lindgren <tony@atomide.com>